@@ -1,17 +1,17 @@
-import { Component, OnInit, Output, Input, EventEmitter } from '@angular/core';
 import { SearchResponse } from 'app/shared/ResourceModels/SearchResponse';
-import { ControlValueAccessor } from '@angular/forms/src/directives/control_value_accessor';
 import { FilteringSearch } from 'app/shared/ResourceModels/FilteringSearch';
 import { ItemDescription } from 'app/shared/ResourceModels/ItemDescription';
 import { SearchService } from 'app/shared/services/search.service';
-
+import { Component, OnInit, Output, Input, EventEmitter } from '@angular/core';
+import { Observable } from 'rxjs/Observable';
+import { map } from 'rxjs/operators';
 @Component({
   selector: 'app-filter-box',
   templateUrl: './filter-box.component.html',
   styleUrls: ['./filter-box.component.scss']
 })
-export class FilterBoxComponent implements OnInit, ControlValueAccessor {
-  @Input() itemDescriptions: ItemDescription[];
+export class FilterBoxComponent implements OnInit {
+  @Input() itemDescriptions: ItemDescription[]= new Array();
   @Input() _selectedValue: string;
   @Output() filteredBy: EventEmitter<any> = new EventEmitter();
   @Output() closeFilterBar: EventEmitter<boolean> = new EventEmitter();
@@ -38,6 +38,16 @@ export class FilterBoxComponent implements OnInit, ControlValueAccessor {
       this.itemDescriptions = x;
       console.log('this.itemDescriptions: ', this.itemDescriptions);
     });
+
+    //   this._searchService.getItemDescription().pipe(
+    //     map(res => res.values)
+    //   )
+    //   .subscribe(res => console.log(res));
+    // }
+
+    this._searchService
+      .getItemDescription()
+      .subscribe(res => console.log(res));
   }
 
   changedFilterBy(event: { target; value: string }, field: string) {
